\name{Bayesianbetareg-package}
\alias{Bayesianbetareg-package}
\docType{package}
\title{Bayesian beta regression package: joint mean and precision modeling}
\description{Bayesian beta regression package}
\details{
\tabular{ll}{
Package: \tab Bayesianbetareg\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2013-08-31\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
 Daniel Jaimes, Margarita Marin, Javier Rojas, Martha Corrales,  Maria Fernanda Zarate,  
 Ricardo Duplat,  Luis Villaraga, Edilberto Cepeda-Cuervo.
}
\keyword{package}




