\name{summary.Bayesianbetareg}
\alias{summary.Bayesianbetareg}
\title{Print the Bayesian beta regression}
\description{Summarized the Bayesian Beta Regression for joint modelling of mean and variance}
\usage{
\method{summary}{Bayesianbetareg}(object, \ldots)
}
\arguments{
  \item{object}{an object of class Bayesianbetareg}
   \item{\ldots}{not used.}
}
\value{
\item{call}{Call}
\item{coefficients}{Coefficients}
\item{Deviance}{Deviance}
\item{AIC}{AIC}
\item{BIC}{BIC}
}
\references{
1.Cepeda C. E. (2001). Modelagem da variabilidade em modelos lineares generalizados. Unpublished Ph.D. tesis. Instituto de Matemáticas. Universidade Federal do Río do Janeiro.http://www.docentes.unal.edu.co/ecepedac/docs/MODELAGEMDAVARIABILIDADE.pdf
2.Cepeda, E. C. and Gamerman D. (2005). Bayesian Methodology for modeling parameters in the two-parameter exponential family. Estadistica 57, 93 105.
3.Cepeda, E. and Garrido, L. (2011). Bayesian beta regression models: joint mean and precision modeling. Universidad Nacional
4.Cepeda, E. and Migon, H. and Garrido, L. and Achcar, J. (2012) Generalized Linear models with random effects in the two-parameter exponential family. Journal of Statistical Computation and Simulation. 1, 1 13
}
\author{
 Daniel Jaimes \email{dajaimesc@unal.edu.co},
 Margarita Marin \email{mmarinj@unal.edu.co},
 Javier Rojas \email{jarojasag@unal.edu.co},
 Martha Corrales \email{martha.corrales@usa.edu.co},  
 Maria Fernanda Zarate \email{mfzaratej@unal.edu.co},  
 Ricardo Duplat \email{rrduplatd@unal.edu.co},    
 Luis Villaraga \email{lfvillarragap@unal.edu.co},
 Edilberto Cepeda Cuervo \email{ecepedac@unal.edu.co}        
}

\keyword{Bayesian}
\keyword{Metropolis Hastings}
\keyword{Mean covariance modelling}
\keyword{Beta}
