\name{Bayesianbetareg-package}
\alias{Bayesianbetareg-package}
\docType{package}
\title{Bayesian beta regression package: joint mean and precision modeling}
\description{Bayesian beta regression package}
\details{
\tabular{ll}{
Package: \tab Bayesianbetareg\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2013-08-31\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
 Daniel Jaimes \email{dajaimesc@unal.edu.co},
 Margarita Marin \email{mmarinj@unal.edu.co},
 Javier Rojas \email{jarojasag@unal.edu.co},
 Hugo Andres Gutierrez Rojas \email{hugogutierrez@usantotomas.edu.co},
 Martha Corrales \email{martha.corrales@usa.edu.co},  
 Maria Fernanda Zarate \email{mfzaratej@unal.edu.co},   
 Ricardo Duplat \email{rrduplatd@unal.edu.co},    
 Luis Villaraga \email{lfvillarragap@unal.edu.co}, 
 Edilberto Cepeda-Cuervo \email{ecepedac@unal.edu.co}        
}
\keyword{package}




